Add macOS testing to CI (#57)
Platform coverage in CI is valuable for C++ since the toolchains are so
variable. This PR should ensure nobody accidentally breaks macOS
support.

Windows does not currently work due to cel-cpp issues. I don't think we
can put any priority on patching that right now. I think MSVC testing
would go a long way to attest that changes are more properly portable,
versus GCC and Clang which are fairly similar. Hopefully, we can
eventually rectify this. The necessary workflow change is already
present in this PR, so once `cel-cpp` is fixed we just need to add
`windows-latest` to the matrix. Some day.

This PR is one I worked on a long time ago, but got stalled on. I
rebased it and cleaned it up since I think it is valuable enough to
merge.