Add field and rule value to violations (#64)
Adds the ability to access the captured rule and field value from a
Violation.

C++ protobuf doesn't really have a native "value" representation unlike
other languages, so I've created a sort-of stand-in. The `ProtoField`
class acts as a reference to a field on a message, and provides an
interface to get a singular value as a `variant`. This should be good
enough for most users, but users can also grab the information needed
and use protobuf reflection directly instead if they have more
complicated needs.

**This is a breaking change.** The API changes in the following ways:
- `Validator::Validate()` now returns the wrapper type
`ValidationResult` instead of `Violations`. This can be converted into
the protobuf `Violations` form using the `proto()` method.
- `ValidationResult.violations()` returns a `ConstraintViolation`. The
underlying `Violation` can be accessed using the `proto()` method.