Implement shared dependencies system (#70)
In preparation for CMake support (#54), we need a system to share
dependencies across build systems. To do this, a technique is needed
which can work with both bzlmod and WORKSPACE mode, as well as CMake.

I have chosen to specify the dependency information in a JSON manifest.
This manifest contains the necessary information to derive the sources
for dependencies that are shared between build systems. For dependencies
which are only needed for Bazel, such as rules repositories or
dependencies of them, those will be specified in `deps.bzl` still. I
don't think we will need any CMake-specific dependencies.

We may need to modify/extend this later for bzlmod support.