Fixes for hostname, port, and enum validation
The previous PRs introduce unified validation for a couple of particularly challenging cases (IPv4/IPv6 addresses and URIs.) This PR is relatively light, since the existing validation functions in protovalidate-cc are already structured close to what they need to be.

For enum validation, all that was wrong was the message.

There are two more classes of conformance test failures:
- Bugs in the CEL string.format helper.
- A bug in v0.10.3 of the conformance test suite that set IGNORE_ALWAYS on some fields that shouldn't have it.

They will be addressed in future PRs, which should allow us to empty out expected_failures.yaml again.