Simplify IPv6 parsing logic a little (#85)
In order to work around some edge cases with IPv6, I adopted a local
state value that would ensure productions could only appear in specific
states. However, the same problem is solved a bit more elegantly in
https://github.com/bufbuild/protovalidate-go/pull/215 by simply always
ensuring that the state after successfully consuming a production can't
be invalid. This means that we need to handle the following edge cases:
- Before, we check that the dotted IP follows a separator or double
colon; now we just check that the dotted IP production does not begin on
a colon. This works because as long as there is no colon at the
beginning, if we pass the dotted check, the only possible valid
production is a dotted IP, and further checks prevent us from invalidly
reaching this state without a separator (because the hexadecatet
production would eat into the dotted production if there was no
separator.)
- Before, we check that a hexadecatet doesn't follow another
hexadecatet; this was not necessary since `parseHexadecimalHexadecatet`
already ensures that the production is not overly long and fails if it
is.
- Before, we check that double colons don't follow a separator. This was
not necessary because double colons are already consumed at a higher
priority, so this state is impossible to reach.
- Before, we check that a colon only follows a hexadecatet. We can now
rule this out by ensuring that a colon doesn't follow a double colon,
and that a single colon is only encountered after at least one
hexadecatet and never at the end of the string. This winds up being
effectively equivalent as those are the only two cases you can possibly
still reach this production that are invalid.
- Before, we check that the parser ends on either a double colon or a
hexadecatet. This rules out ending on an empty string, which is not
necessary: that would imply zero hexadecatets were parsed, which will
return false anyways shortly after. This also rules out ending on a
colon, which we can enforce by ensuring that the input string is not
empty before continuing.

So I believe this is *at least* as correct as it was before, but is
simpler.