Update cel-cpp to v0.11.0
This is a bit complicated because cel-cpp v0.11.0 breaks WORKSPACE support a little bit, but thankfully the API we're using has remained relatively stable. Patches have been rebased, I'm pretty confident that I'll need to loop around to the MSVC patch to make sure it still works but the field access patch seems to be working first-try so knock-on-wood. The CMakeLists will need to be audited because I'm sure there's now stuff in it that should be removed, but let's focus on getting a working build first. We can focus on making the CMakeLists as accurate as possible when we're closer to being able to upstream it.