New conformance fixes
Removes the string format library in favor of cel-cpp's standard implementation. This causes some tests to be fixed, but it also causes some new failures due to non-standard string formatting behavior in protovalidate. This will be fixed on the next protovalidate release, so we should probably release that and then update this PR with it before merging.