Attempt to make SLSA work again
It seems to not like an explicit commit ref... will it accept a branch ref instead? I tried looking at the SLSA source code to find out, but it's going to be easier to just try. If not I need to bug them to make a new release because a change we need is not tagged yet.