Use tagged BCR release workflow
While waiting for the changes that allow the BCR release workflow to create a draft release instead of publishing the release immediately, I was using a branch reference to try to get SLSA to work. Unfortunately that doesn't work: the SLSA attestation won't be trusted unless we're using a tagged version of the workflow. At some point, a tag was created for the merged change, so we can move to it now. With any luck, this should allow the next release of protovalidate-cc to pass attestation.